Contact page update: changed embedded Google Map iframe source; added margin to radio buttons in contact form
In the contact.html file, the embedded Google Map iframe source has been changed to a different URL. Additionally, in the same file, a margin has been added to the radio buttons in the contact form by wrapping them in a div with the class "radio-container" and adding CSS to create the margin.

In the rafting.css file, the width property has been added to the .map-container iframe selector with a value of 400.

Note: There is an extra "br" tag in the contact.html file that was not part of the original diff and can be disregarded.