Contact page update: changed embedded Google Map iframe source; added margin to radio buttons in contact form
In the contact.html file, the embedded Google Map iframe source has been changed to a different URL. Additionally, in the same file, a margin has been added to the radio buttons in the contact form by wrapping them in a div with the class "radio-container" and adding a CSS rule for this class in the rafting.css file. A empty <p> and <br> tag have been added to improve readability.