"Added margin to nav, modified specificity of nav a styles"
This commit message describes the changes made to the CSS file for a rafting website. It mentions that a margin has been added to the nav element and that the specificity of the nav a styles has been modified. It is concise and accurately describes the changes made in the code diff.